Moved independent files to respective packages. Removed onCreate method call in fragments and passed layout ids in constructors.
