+++ app/src/main/res/layout/activity_google_pay.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GooglePayActivity">
+    tools:context=".google_pay.GooglePayActivity">
 
     <Button
         android:id="@+id/btn_google_pay"
