+++ app/src/main/res/layout/fragment_get_card_info.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".headless.GetCardInfoFragment">
+    tools:context=".headless.get_card_info.GetCardInfoFragment">
 
     <EditText
         android:id="@+id/et_card_info"
