+++ app/src/main/res/layout/fragment_pay_with_saved_payment_method.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".headless.PayWithSavedPaymentMethod">
+    tools:context=".headless.pay_with_saved_payment_method.PayWithSavedPaymentMethod">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
