+++ app/src/main/res/layout/fragment_pay_with_saved_payment_options.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".headless.PayWithSavedPaymentOptions">
+    tools:context=".headless.pay_with_saved_payment_method.PayWithSavedPaymentOptions">
 
     <TextView
         android:layout_width="match_parent"
