+++ app/src/main/res/layout/fragment_payment_options.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".headless.PaymentOptionsFragment">
+    tools:context=".headless.make_payment.PaymentOptionsFragment">
 
     <TextView
         android:id="@+id/title"
