+++ app/src/main/res/layout/fragment_save_payment_method.xml
@@ -4,7 +4,7 @@
     android:id="@+id/rl_save_payment_method"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".headless.SavePaymentMethod">
+    tools:context=".headless.save_payment_method.SavePaymentMethod">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
