+++ app/src/main/res/layout/fragment_save_payment_method_payment_options.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".headless.SavePaymentMethodPaymentOptions">
+    tools:context=".headless.save_payment_method.SavePaymentMethodPaymentOptions">
 
     <TextView
         android:id="@+id/title"
