+++ app/src/main/res/layout/fragment_validate_fields_payment_options.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".headless.ValidateFieldsPaymentOptions">
+    tools:context=".headless.validate_fields.ValidateFieldsPaymentOptions">
 
     <TextView
         android:id="@+id/title"
